document_generation/requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
httpx
openai
python-dotenv